\name{Bonus-Lab-package}
\alias{Bonus-Lab-package}
\alias{Bonus-Lab}
\docType{package}
\title{
\packageTitle{Bonus-Lab}
}
\description{
\packageDescription{Bonus-Lab}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bonus-Lab}
\packageIndices{Bonus-Lab}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Bonus-Lab}

Maintainer: \packageMaintainer{Bonus-Lab}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
