% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_airport_delays.R
\name{visualize_airport_delays}
\alias{visualize_airport_delays}
\title{Visualize Airport Delays}
\usage{
visualize_airport_delays()
}
\description{
Creates a scatter plot of mean delays by airport location.
}
\details{
This function visualizes mean departure delays for flights to different airports 
         by plotting each airport's longitude and latitude, with the size and color of the points 
         representing the average delay.
}
